move useDeleteVideo to hooks
