+++ src/hooks/index.ts
@@ -10,3 +10,4 @@ export * from './useJoystream'
 export * from './useConnectionStatus'
 export * from './useEditVideoSheet'
 export * from './useUploadsManager'
+export * from './useDeleteVideo'
